bender/requirements.txt
aiohttp==1.0.5
asn1crypto==0.23.0
async-timeout==2.0.0
asyncssh==1.11.1
cffi==1.11.2
chardet==3.0.4
cryptography==2.1.3
discord.py==0.16.12
idna==2.6
multidict==3.3.2
pycparser==2.18
six==1.11.0
websockets==3.4
